+++ graylog2-web-interface/src/components/common/Section/SectionComponent.jsx
