+++ graylog2-web-interface/src/components/users/UserDetails/ProfileSection.jsx
@@ -2,8 +2,8 @@
 import * as React from 'react';
 
 import User from 'logic/users/User';
+import SectionComponent from 'components/common/Section/SectionComponent';
 
-import SectionComponent from '../SectionComponent';
 import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
 import LoggedInIcon from '../LoggedInIcon';
