+++ graylog2-web-interface/src/components/users/UserDetails/SettingsSection.jsx
@@ -2,8 +2,8 @@
 import * as React from 'react';
 
 import User from 'logic/users/User';
+import SectionComponent from 'components/common/Section/SectionComponent';
 
-import SectionComponent from '../SectionComponent';
 import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
 
 type Props = {
