+++ graylog2-web-interface/src/components/users/UserEdit/ProfileSection.jsx
@@ -4,8 +4,8 @@ import { Formik, Form } from 'formik';
 
 import { Button, Col, Row } from 'components/graylog';
 import User from 'logic/users/User';
+import SectionComponent from 'components/common/Section/SectionComponent';
 
-import SectionComponent from '../SectionComponent';
 import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
 import FullNameFormGroup from '../UserCreate/FullNameFormGroup';
 import EmailFormGroup from '../UserCreate/EmailFormGroup';
